do a LOT
I've moved a couple types so that they're in a more sensible place. I've added a few comments. Added a dev command line argument that automatically connects to a mongodb instance at localhost. Add a function to validate an answer to a question. It doesn't actually compare it to the question's answer just checks it's format. Make multiple choice a bit more functional. There's quite a lot more, I need to start committing more often, but this is most of the important stuff I think. Oh, also I added a confirmation prompt that confirms if you want to save the question or not.